[accumulator][fuzz] convert proptest to fuzzers
Closes: #5956
