+++ storage/accumulator/Cargo.toml
@@ -15,6 +15,7 @@ libra-crypto = { path = "../../crypto/crypto", version = "0.1.0" }
 mirai-annotations = "1.10.1"
 libra-types = { path = "../../types", version = "0.1.0" }
 libra-workspace-hack = { path = "../../common/workspace-hack", version = "0.1.0" }
+proptest = { version = "0.10.1", optional = true }
 
 [dev-dependencies]
 rand = "0.7.3"
@@ -24,4 +25,4 @@ libra-crypto = { path = "../../crypto/crypto", version = "0.1.0", features = ["f
 
 [features]
 default = []
-fuzzing = ["libra-crypto/fuzzing", "libra-types/fuzzing"]
+fuzzing = ["proptest", "libra-crypto/fuzzing", "libra-types/fuzzing"]
