+++ testsuite/libra-fuzzer/Cargo.toml
@@ -26,6 +26,7 @@ libra-proptest-helpers = { path = "../../common/proptest-helpers", version = "0.
 libra-workspace-hack = { path = "../../common/workspace-hack", version = "0.1.0" }
 
 # List out modules with data structures being fuzzed here.
+accumulator = { path = "../../storage/accumulator", version = "0.1.0", features = ["fuzzing"] }
 consensus = { path = "../../consensus", version = "0.1.0", features = ["fuzzing"] }
 consensus-types = { path = "../../consensus/consensus-types", version = "0.1.0", features = ["fuzzing"] }
 executor = { path = "../../execution/executor", version = "0.1.0", features = ["fuzzing"] }
