+++ testsuite/libra-fuzzer/src/fuzz_targets/storage.rs
@@ -2,6 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{corpus_from_strategy, fuzz_data_to_value, FuzzTargetImpl};
+use accumulator::test_helpers::{
+    arb_hash_batch, arb_list_of_hash_batches, arb_three_hash_batches, arb_two_hash_batches,
+    test_append_empty_impl, test_append_many_impl, test_consistency_proof_impl,
+    test_get_frozen_subtree_hashes_impl, test_proof_impl, test_range_proof_impl,
+};
 use libra_jellyfish_merkle::test_helper::{
     arb_existent_kvs_and_nonexistent_keys, arb_kv_pair_with_distinct_last_nibble,
     arb_tree_with_index, test_get_range_proof, test_get_with_proof,
@@ -106,5 +111,115 @@ impl FuzzTargetImpl for JellyfishGetRangeProof {
     fn fuzz(&self, data: &[u8]) {
         let input = fuzz_data_to_value(data, arb_tree_with_index(1000));
         test_get_range_proof(input);
+    }
+}
+
+#[derive(Clone, Debug, Default)]
+pub struct AccumulatorFrozenSubtreeHashes;
+
+impl FuzzTargetImpl for AccumulatorFrozenSubtreeHashes {
+    fn description(&self) -> &'static str {
+        "Accumulator frozen subtree hashes"
+    }
+
+    fn generate(&self, _idx: usize, _gen: &mut ValueGenerator) -> Option<Vec<u8>> {
+        Some(corpus_from_strategy(arb_hash_batch(1000)))
+    }
+
+    fn fuzz(&self, data: &[u8]) {
+        let input = fuzz_data_to_value(data, arb_hash_batch(1000));
+        test_get_frozen_subtree_hashes_impl(input);
+    }
+}
+
+//============== Accumulator =============
+
+#[derive(Clone, Debug, Default)]
+pub struct AccumulatorProof;
+
+impl FuzzTargetImpl for AccumulatorProof {
+    fn description(&self) -> &'static str {
+        "Accumulator proof"
+    }
+
+    fn generate(&self, _idx: usize, _gen: &mut ValueGenerator) -> Option<Vec<u8>> {
+        Some(corpus_from_strategy(arb_two_hash_batches(100)))
+    }
+
+    fn fuzz(&self, data: &[u8]) {
+        let input = fuzz_data_to_value(data, arb_two_hash_batches(100));
+        test_proof_impl(input);
+    }
+}
+
+#[derive(Clone, Debug, Default)]
+pub struct AccumulatorConsistencyProof;
+
+impl FuzzTargetImpl for AccumulatorConsistencyProof {
+    fn description(&self) -> &'static str {
+        "Accumulator consistency proof"
+    }
+
+    fn generate(&self, _idx: usize, _gen: &mut ValueGenerator) -> Option<Vec<u8>> {
+        Some(corpus_from_strategy(arb_two_hash_batches(100)))
+    }
+
+    fn fuzz(&self, data: &[u8]) {
+        let input = fuzz_data_to_value(data, arb_two_hash_batches(100));
+        test_consistency_proof_impl(input);
+    }
+}
+
+#[derive(Clone, Debug, Default)]
+pub struct AccumulatorRangeProof;
+
+impl FuzzTargetImpl for AccumulatorRangeProof {
+    fn description(&self) -> &'static str {
+        "Accumulator range proof"
+    }
+
+    fn generate(&self, _idx: usize, _gen: &mut ValueGenerator) -> Option<Vec<u8>> {
+        Some(corpus_from_strategy(arb_three_hash_batches(100)))
+    }
+
+    fn fuzz(&self, data: &[u8]) {
+        let input = fuzz_data_to_value(data, arb_three_hash_batches(100));
+        test_range_proof_impl(input);
+    }
+}
+
+#[derive(Clone, Debug, Default)]
+pub struct AccumulatorAppendMany;
+
+impl FuzzTargetImpl for AccumulatorAppendMany {
+    fn description(&self) -> &'static str {
+        "Accumulator amend many leaves"
+    }
+
+    fn generate(&self, _idx: usize, _gen: &mut ValueGenerator) -> Option<Vec<u8>> {
+        Some(corpus_from_strategy(arb_list_of_hash_batches(10, 10)))
+    }
+
+    fn fuzz(&self, data: &[u8]) {
+        let input = fuzz_data_to_value(data, arb_list_of_hash_batches(10, 10));
+        test_append_many_impl(input);
+    }
+}
+
+#[derive(Clone, Debug, Default)]
+pub struct AccumulatorAppendEmpty;
+
+impl FuzzTargetImpl for AccumulatorAppendEmpty {
+    fn description(&self) -> &'static str {
+        "Accumulator amend empty leave"
+    }
+
+    fn generate(&self, _idx: usize, _gen: &mut ValueGenerator) -> Option<Vec<u8>> {
+        Some(corpus_from_strategy(arb_hash_batch(100)))
+    }
+
+    fn fuzz(&self, data: &[u8]) {
+        let input = fuzz_data_to_value(data, arb_hash_batch(100));
+        test_append_empty_impl(input);
     }
 }
